Move enum to a separate class
