+++ server/src/test/java/com/objectcomputing/checkins/services/member_skill/skillsreport/SkillsReportServicesImplTest.java
@@ -59,7 +59,7 @@ public class SkillsReportServicesImplTest {
     void testReportSkillNotExist() {
         final SkillLevelDTO dto = new SkillLevelDTO();
         dto.setId(UUID.randomUUID());
-        dto.setLevel(SkillsReportServicesImpl.SkillLevel.INTERMEDIATE);
+        dto.setLevel(SkillLevel.INTERMEDIATE);
 
         final List<SkillLevelDTO> skills = new ArrayList<>();
         skills.add(dto);
@@ -155,10 +155,10 @@ public class SkillsReportServicesImplTest {
         final SkillLevelDTO dto2 = new SkillLevelDTO();
         final SkillLevelDTO dto3 = new SkillLevelDTO();
         dto1.setId(skillId1);
-        dto1.setLevel(SkillsReportServicesImpl.SkillLevel.INTERMEDIATE);
+        dto1.setLevel(SkillLevel.INTERMEDIATE);
         dto2.setId(skillId2);
         dto3.setId(skillId3);
-        dto3.setLevel(SkillsReportServicesImpl.SkillLevel.ADVANCED);
+        dto3.setLevel(SkillLevel.ADVANCED);
 
         final List<SkillLevelDTO> requestedSkills1 = new ArrayList<>();
         requestedSkills1.add(dto1);
@@ -222,9 +222,9 @@ public class SkillsReportServicesImplTest {
         final SkillLevelDTO dto4 = new SkillLevelDTO();
         final SkillLevelDTO dto5 = new SkillLevelDTO();
         dto4.setId(skillId2);
-        dto4.setLevel(SkillsReportServicesImpl.SkillLevel.INTERMEDIATE);
+        dto4.setLevel(SkillLevel.INTERMEDIATE);
         dto5.setId(skillId4);
-        dto5.setLevel(SkillsReportServicesImpl.SkillLevel.ADVANCED);
+        dto5.setLevel(SkillLevel.ADVANCED);
 
         final List<SkillLevelDTO> requestedSkills2 = new ArrayList<>();
         requestedSkills2.add(dto4);
@@ -242,9 +242,9 @@ public class SkillsReportServicesImplTest {
         for (SkillLevelDTO skill : response4.getTeamMembers().get(0).getSkills()) {
             assertTrue(skill.getId().equals(skillId2) || skill.getId().equals(skillId4));
             if (skill.getId().equals(skillId2)) {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("intermediate"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("intermediate"), skill.getLevel());
             } else {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("expert"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("expert"), skill.getLevel());
             }
         }
         verify(memberSkillRepository, times(11)).findBySkillid(any(UUID.class));
@@ -259,9 +259,9 @@ public class SkillsReportServicesImplTest {
         for (SkillLevelDTO skill : elem.getSkills()) {
             assertTrue(skill.getId().equals(skillId1) || skill.getId().equals(skillId2));
             if (skill.getId().equals(skillId1)) {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("intermediate"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("intermediate"), skill.getLevel());
             } else {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("advanced"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("advanced"), skill.getLevel());
             }
         }
     }
@@ -272,9 +272,9 @@ public class SkillsReportServicesImplTest {
         for (SkillLevelDTO skill : elem.getSkills()) {
             assertTrue(skill.getId().equals(skillId2) || skill.getId().equals(skillId3));
             if (skill.getId().equals(skillId2)) {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("interested"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("interested"), skill.getLevel());
             } else {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("advanced"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("advanced"), skill.getLevel());
             }
         }
     }
@@ -285,9 +285,9 @@ public class SkillsReportServicesImplTest {
         for (SkillLevelDTO skill : elem.getSkills()) {
             assertTrue(skill.getId().equals(skillId1) || skill.getId().equals(skillId2));
             if (skill.getId().equals(skillId1)) {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("advanced"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("advanced"), skill.getLevel());
             } else {
-                assertEquals(SkillsReportServicesImpl.SkillLevel.convertFromString("intermediate"), skill.getLevel());
+                assertEquals(SkillLevel.convertFromString("intermediate"), skill.getLevel());
             }
         }
     }
